requirements.txt
asgiref==3.9.1
bcrypt==4.2.0
blinker==1.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
filelock==3.16.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.1.1
packaging==24.2
pandas==2.2.3
pipenv==2024.4.0
platformdirs==4.3.6
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
Werkzeug==3.0.1